Workaround for #50
This is terrible. Cabal is half-broken: in some places it correctly
ignores buildable:False stanzas. In other places, it doesn't. Chris was
using one of the non-broken pieces of the Cabal code base, which meant
that we were (correctly) getting a smaller dependency set for some
packages. Unfortunately, this didn't jive with Cabal's broken view of
the world, which decided we didn't have all necessary dependencies
installed.

Solution: copy the (correctly implemeneted) allBuildInfo function from
Cabal and break it in the same way that the other parts of Cabal are
broken. I vomited a little in my mouth while typing that.