More explicit EnvOverride (fixes #57)
Turned out to be a bit more involved than I'd hoped, but the changes are
all mechanical. Force more thought into the environment variables for
subprocesses by making it an explicit parameter. Then in the Config,
have a function for generating this environment, which lets us control
whether local databases are included and whether we set GHC_PACKAGE_PATH