Concurrently stream to file for large stderr
Additionally: fix logging from runhaskell processes

The main bug was that both the stdout and stderr logging was trying to
open up the same file, and only one of them would succeed. However, the
function was swallowing all exceptions, so this was non-obvious.
Previously, this wasn't an issue because the two streams were logged
sequentially, but that needed to change for the case of large output.

Once I removed the swallowed exception, the bug became obvious, and now
the code opens up a Handle that is shared by both stdout and stderr.

Without this fix, turning on the multithreaded runtime causes surprising
exceptions because the Handles from the child process end up getting
closed nondeterministically, which causes runhaskell Setup.hs to fail
with "resource vanished."