Use a stack-specific dist dir #128
This affected a lot of code since I needed to change some .Constants functions
to (1) take a PackageIdentifier for the Cabal version (more on that in a sec)
and (2) live in MonadThrow. I'm somewhat tempted to get rid of the second and
make the function "partial," based on the knowledge that rendering a version
will always generated a valid relative dir. @chrisdone Thoughts on that?

As for the Cabal version: this is intended to work around a limitation in Cabal
that resulted in a bug (#69). By putting all files in a directory with the
Cabal version in the name, we should avoid Cabal getting confused by
"inappropriate type" of files.