Make more builds "singular"
I was trying to run the test suite for warp and, to my surprise, the
output was not being sent to the console but instead to a log file. I
realized that under the previous code, the mere presence of extra
packages in the stack.yaml file would prevent console output. However,
we'd commonly want that output anyway, such as my case of running `stack
test` within the warp directory.