Read/Run process running distinction
@manny-fp @snoyberg @danburton There's now System.Process.Read and
System.Process.Run; the former is for consuming the output from
processes in a way that doesn't touch the current process's
stdout/stderr. The latter is for running a process and having the stdout
and/or stderr go to the current process's. Please try to keep functions
added to these modules named like "read" or "run" so it's easy to know
what the intention is.