Disable retrying strategy on filelock.  Instead use a blocking call after first fail.
Originally, I had intendeded to add some more warning messages for feedback
if it failed to acquire the lock for too long.  The retry/backoff strategy
should not have used too much CPU to poll the lock a couple of times a second.
However, on my laptop I caught it spiking to 100% CPU and don't know why.