Place cabal constraints in config file, not args #723
It seems that, on Windows, the large number of arguments causes the process
package to give the incredibly misleading error message:

C:\Users\chris\AppData\Roaming\local\bin\cabal.exe: streamingProcess: does not exist (No such file or directory)

This workaround isn't just a hack, but a good thing: putting constraints in the
config file is an improvement in general, so I'm happy to keep it. Nonetheless,
this is something that should be fixed at the process package level.