add typo recommendations, close #158
This commit introduces typo recommendations and minor refactorings in
‘Stack.Fetch’ module. Currently the recommendations follow very
conservative scheme only reporting package names for which
Damerau-Levenshtein distance between given package and package name in
caches is equal to 1. This should catch all common typos. If desirable,
the threshold can be made greater in future, although do it care because
it will quickly start to give false positives.