Fix building tests/benches in some circumstances
It was doing the "all-in-one" build logic when initially adding wanted
dependencies. However, this broke in the circumstance where some other
local package depends on the package with the tests / benchmarks.  If
the dependencies for the downstream package are chased first, this led
to the "all-in-one" logic not executing.

Regression introduced in 85c5a3973a0a7199cefe522ab820a070c0961440