Implement executable fall-through
This closes #1433. When an invalid `stack` command is issued, `stack`
will look in the `PATH` for a matching binary. So given there's an
executable `stack-something` in the `PATH`, issuing `stack something`
will execute it, rather than failing.

It uses `System.Directory.findExecutable` to find executables and `exec`
from `System.Process.Read` to execute the process.

Adds a mOnFailure callback to be called on parser failures to the
`Options.Complicated` parsing function.