Split Data/Attoparsec/Args.hs
Using the constant stackProgName in Args.hs creates a import cycle. Split the
file in two parts to avoid that. The interpreter argument parsing which
requires the constant is moved to a separate module.

This change only splits the file with no functionality change at all to
facilitate easier review of the changes.