Add zsh completion function
The files _completion_helpers and _stack provide much better completion
than the bash completer. They provide command and options descriptions, and
completion of subcommands and option arguments.
The completion function is dynamically generated based on --help output
for stack and it's subcommands, so that it should always be up to date
as long as the help output keeps the same format.