Redownload files with incorrect hash
Previously, if a downloaded file had the right size but the wrong hash,
Stack would simply exit without redownloading. The problem is simply one
of parentheses: we need the catch statement to apply to the action
itself, not the first exception handler.

Alternative solution (arguably better): combine the two exception types
into one.