Track the best snapshot by the number of deperrors
Changes in this commit:

1) When figuring out a build plan, track the dependency errors for each
snapshot to figure out which snapshot is the best even if there are
dependency errors.

2) Use the same strategy even when finding out the best set of flags.
Find out the combination of flags producing least number of dependency
errors.

3) Modularise the interfaces to make the code components more reusable
if needed.

This change is supposed to be used for later enhancements to use the
dependency solver on the best matching snapshot in absence of an error
free match.