Track dependency errors for GHC wired in packages
This commit adds ability to track and distinguish build plan dependency errors
related to the packages which are wired in GHC. This interface will be used to
detect whether a snapshot can be used at all for building a given package. If
GHC wired in packages are incompatible we cannot use the snapshot even with
extra dependencies.

This is not being used yet but will be used in future commits.