Reverse dependency between Init and Config modules
Need to import Stack.Setup in Stack.Solver to be able to use ensureCompiler
which creates a dependency cycle. The cycle can be fixed by making Init
depend on Config instead of the opposite.

This commit only includes the code movement due to the above. There are
no functionality changes.