Ensure, use specified compiler for new/init/solver
This commit introduces the following enahncements:

1) Support --install-ghc for stack new/init/solver commands

2) init or solver will now use the correct compiler specified
in the resolver on command line during init or in stack.yaml when using
the solver command.

Before this change solver used the compiler found on path disregarding
any configuration settings (for stack solver) or the snapshot chosen
(for stack init) or the resolver specified on command line.

Note that even with this change stack init --solver still returns a
compiler resolver with all deps as extra-deps. Using a snapshot resolver
with minimal extra-deps is a matter of a future commit.

Fixes #1529