Solver: Use snapshot instead of compiler resolver
Changes include:

1) For 'init --solver' use the best selected snapshot (having minumum number
of dep errors) with minimum extra deps on top of the snapshot. Also use
the auto selected flags producing minimal snapshot dep errors for solving.

Before this change, solver always used a compiler resolver and
everything as extra-deps.

2) Fix a bug in solver. When determining new extra-deps take version
number into account, not just the name. If the version is different
from the previous or in-snapshot version then its a new dependency.