Solver: retry with relaxed contraints if failed
With a recent commit solver was made to use packages in the snapshot as
contraints when solving. Within those constraints it tries to figure what
external packages are needed to build the package. But this strategy may not
always work. The package dependencies may want to use a package which is
already in the snapshot but requires a version different than in the snapshot.

To take care of this we first try with the snapshot packages as hard
constraints and if that fails we try solving with those packages as soft
constraints i.e. merely preferences.

This commit implements this fallback strategy. With this in place we should be
able to solve and init any package which can be solved by cabal. If any package
is not solved then either the package is broken or we have a bug in stack
solver.

This commit also added useful information in the status logs and suggestions
for next steps or resolutions in case solver fails.