Solver: perform compiler compatibility check
With recent changes we perform compiler compatibilty check when choosing
snapshots for stack init. We need similar checks for stack solver as well.

This change uses the same code to do compiler compatibility check when using
the solver command. If the resolver specified in the stack.yaml does not have a
compiler which can work with the package we emit an appropriate messages
instead of going ahead and then showing difficult to understand cabal errors.