Take flags into account to compute extra deps
Solver issues fixed with this commit are:

When solver returns flag changes for packages which are in snapshot, we update
the flags in stack.yaml but we do not put those packages in extra dependencies.
When we run stack build it complains about it.

We are not comparing the snapshot package flags (we compare only the version)
when determining extra dependencies. If the snapshot flags are different than
those the solver arrived at then we need to put those packages in extra
dependencies. Otherwise the plan will not build because the status of flags can
change dependencies.