Show flags along with dep errors in debug mode
The build plan checks show dependency errors about packages when a particular
build plan does not match. But flags can affect dependencies and are therefore
needed to verify the errors. This commit adds code to print the flags as well
in debug mode. Avoid in normal mode for the sake of brevity of info.