stack init: ignore resolver incompatible packages
When doing stack init, some of the packages may not be compatible with the
resolver compiler while others just build ok. In that case we provide the user
an option to ignore those packages and use the rest. If --force is
specified stack will create a config with those packages which are compatible.

Fixes #1621