Solver: remove one package in conflict and retry
When there is an unresolved conflict among the dependencies of multiple source
packages then remove one of the conflicting packages and then retry. The
package chosen to be removed is the one which is on top of the dependency
pyramid i.e. noone else depends on it.

The functionality is not yet complete. It will be complete once cabal output is
parsed and the list of conflicting packages is fed to the upper level logic.

See #1616