Solver: resolve conflict by ignoring packages
When there is an unresolvable conflict among the dependencies of multiple
user packages ignore one of them and try again to resolve the rest.

This commit adds code for parsing cabal output to find out user packages
involved in the conflict. This output is then used to decide one of those
packages to be ignored from consideration in the next try.

Fixes #1616