Add 'allow-different-user' flag and configuration option #471, fix #1777
Users other than the owner of the ~/.stack directory are now prevented
from using a stack installation in order to avoid problems with file
permissions. To disable this precaution users can pass the
--allow-different-user flag or use the 'allow-different-user'
configuration option in their ~/.stack/config.yaml.

On Windows, the new flag and configuration options have no effect.

When stack is re-spawned in Docker, the ownership check is skipped
(see #1777).

This reverts commits 0a89c9c29d744a4955e62b59774036e98d6b1e8e and
218e7ddb90c346d814474cb6b8c5e9c67c4706f8.