Revert "init: check pkg name and .cabal file name match"
This reverts commit 63c7f48f6b6e7387e7ffedf5bd480114d37aff2e.

@harendra-kumar The reverted commit seems to break the
1336-1337-new-package-names integration test on Mac OS X.  I did
a bit of debugging and it appears that, somewhere along the way,
some unicode characters are changing (they look the same when
rendered, but the UTF-8 bytes are different) and that is causing
the comparison to fail.

For example:

```
$ stack new ば日本-4本
[...]
Package name as defined in the .cabal file must match the .cabal file name.
Please fix the following packages and try again:
- ば日本-4本/ば日本-4本.cabal
```

`show (FP.takeBaseName $ toFilePath fp,show $ gpdPackageName gpd)`
in this case gives me
`("\12399\12441\26085\26412-4\26412","\12400\26085\26412-4\26412")`;
note the first two bytes in the fst are different than the first byte
in the snd.