Improve file and directory resolution
Previously the idiom ‘forgivingAbsence (relsoveFile …)’ alone was used,
which relied on ‘canonicalizePath’ throwing ‘isDoesNotExistError’ when
path does not exist. As it turns out, this behavior is actually not
intentional and unreliable, see

https://github.com/haskell/directory/issues/44

This was “fixed” in version 1.2.3.0 of ‘directory’ package (now it never
throws). To make it work with all versions, we need to use the following
idiom:

forgivingAbsence (resolveFile …) >>= rejectMissingFile

This commit fixes all cases where new directory versions might cause
problems.