Avoid expensive version calculations during build plan construction
This change delays searching the package caches for versions to
the point where the versions are actually used. As the versions
are only used for error messages in the case of version conflicts,
this change results in a noticeable speedup for no-op builds.

While this change does not remove the call to getPackageCaches as
originally intended, the call should not have any performance impact
as the package caches have already been loaded and cached in the
withLoadPackage function that wraps constructPlan.