Allow global Stack arguments with a script
Stack has support for running Haskell scripts directly, when you pass a
filename as an argument.

Previously this worked by having a fallback: if any argument couldn't be
parsed as a valid option, check if the first argument was a file.

Sadly this broke for cases such as:

    stack --nix example.hs

And this means something different:

    stack example.hs --nix

Now Stack will split the arguments into two:

1. All arguments until the first existing file
2. The first existing file and all arguments afterwards

So the above example gets parsed as:

1. `["--nix"]`
2. `["example.hs"]`

The first section will be reparsed as options. The second section will
be used like it was previously.

This fixes the specific case of running interpreter scripts when having
the Nix option enabled globally. Stack would correctly start a
nix-shell. A nix-shell respawns Stack with some extra global
arguments (extra-lib-dirs and extra-include-dirs) but Stack would treat
them as file arguments, because just one of the arguments (the script)
was a file.