Fix use of deprecated cabal flag (fixes #2350)
From cabal 1.21.1 onwards, the flag --enable-profiling is preferred over
--enable-executable-profiling. Cabal emits a warning when using the
latter flag which also happens to be visible to the user, possibly
causing confusion as to whether the warning is due to stack or because
of misconfiguration.

You can reproduce this bug by running stack build --profile or an
equivalent command.

This commit fixes this by using --enable-profiling for cabal 1.21.1
or later.

This fixes #2350.