cabal: Add logic for handling old-locale
Stack requires some TimeLocale stuff, which was added to time in 1.5.
For prior versions of time (i.e. the 1.4 series), we additionally need
the old-locale package, which then provides the locale logic instead of
the time package. This can be easily achieved using a cabal flag.

This is required to build stack on lts-2.