Replace runghc with ghc -e Main.main
Motivation: if you call `runghc`, that process continues to live on
separate from the actual `ghc` process that interprets the code. That's
inefficient. But more worryingly: the `runghc` process itself can cause
major problems in Docker images, since it ends up swallowing SIGINTs as
the PID-1 process. This was discussed quite a bit on Twitter:

https://twitter.com/argumatronic/status/763418990982991872