Warn on missing extra-dep for external packages.
Motivation: the vast majority of the time, an external package should be
treated as an external dependency, in that we don't want to run test
suites. I regularly see user confusion on this topic. It was probably a
mistake on my part to make the default be extra-dep: false for external
packages.

Changing the default would be a breaking change, so instead, this change
will give the user a warning when they use an external package - such as
from a Git repo - without an explicit extra-dep setting. If the user
really wants the current default behavior, they can add `extra-dep:
false` to silence the warning. I expect in most cases users will end up
adding `extra-dep: true` and avoid a stumbling block.