Switch to getGlobalManager #2359
Related to snoyberg/http-client#214. The fixes to improve latency in
creating a manager may give the biggest improvement to the global
manager, so switching over to using that makes sense. As it stands,
Stack is not modifying the manager configuration at all, so this
shouldn't be a problem, and if that changes in the future, we can always
modify the global manager directly with setGlobalManager.

A more invasive change is to remove the Manager from the configuration
environments entirely and either (1) use getGlobalManager throughout the
codebase or (2) switch over to the Network.HTTP.Simple module. I'm not
opposed to doing that necessarily, but wanted to kick off with the
simpler change.