Do not special-case programs directory on Windows
Just use $STACK_ROOT/programs, like on other platforms. This enables the
directory to be configurable by setting $STACK_ROOT, and it removes the
hard-coded path. When downloading GHC, the archive is downloaded to this
directory as well before extraction. (This was the case anyway.) This is
perhaps less idiomatic, but $STACK_ROOT exists on Windows regardless, so
it seems better to keep everything in one place.

Fixes #1644.