Always launch Docker command from the project root
This commit makes it so docker create is always launched with its working
directory set to the project root. That means relative paths in the docker
run-args field work regardless of which directory stack is launched from.
Processes inside the container still run in the working directory stack was
launched from.

Fixes #2757.