Remove verbatim linebreak
Text inside `` is rendered into <code> tags in HTML.  The CSS style
for <code> tags in the resulting page appears to understand line
breaks in the inner HTML of the tag as being verbatim.  Hence it is
probably wise not to break literal expressions in the markdown source
over multiple lines.