Fix HLint/CPP failure in Travis
HLint was failing because it was using --cpp-simple, which just strips
every directive, which in turn created duplicate import statements. This
commit turns off --cpp-simple and lints with the WINDOWS macro on and off.
--cpp-simple is kept on for the tests, which don't have the duplicate import
problem and some of which use CPP in more complex ways than just turning a
single variable on and off.