Avoid missing SHA warning when caching build plan
Previously, the Stack.BuildPlan code was accidentally reintroducing the
originally Git SHA value when calling withCabalFiles. This seems
innocent enough, until you realize that the presence of a Git SHA in
that argument tells withCabalFiles that it was unable to previously look
up the cabal file by SHA. Perhaps that logic should be changed a bit to
be more obvious...

That said, this problem was especially difficult to diagnose since there
were in fact _two_ git SHAs present in the build plan construction: one
telling withCabalFiles which cabal file to look up, and one which would
ultimately be embedded in the cache. I found this too confusing to work
with and too easy to mess up.

Instead, I made things a bit more type safe: instead of some arbitrary
tuples, just reuse the ResolvedPackage type in a few more places.