Revert "Set NoBuffering mode io, before exec #2702"
This reverts commit 6c876293a4ca215c2645de162aa0314f415fb754.

Running hSetBuffering just before exec had no effect on any actual
buffering, but it screwed up the persistent terminal state in such a
way that the runtime fails to reset it.  The effect was that keys like
Backspace, Ctrl-C, Ctrl-D, etc. were not working correctly in programs
launched inside ‘stack exec’ and after ‘stack exec’.

Closes: #2884.

Reopens: #2702 (actually, that was never fixed in the first place).

Signed-off-by: Anders Kaseorg <andersk@mit.edu>