Reword message shown when binary upgrade is skipped
This seems a bit more accurate: When the user is already running the most recent version, the message "your version is already *more* recent" is incorrect and ever so slightly confusing. The other case -- where the user is running a more recent version than is available as a binary, for which the old message is accurate -- likely occurs much less frequently.