Remove a bunch of unneeded info from BuildPlan
This came in automatically at the inception of Stack, by copying data
types from stackage-curator. In reality, we don't need all of this
information within Stack. As we move towards extensible snapshots, we
want to make sure we have the bare minimum of information to allow a
shared data type for parsing all kinds of snapshots.

Additionally, we probably want to move away from depending on extra
information present in the build plan, in case there are mistakes in it.