Do not store core package info in build plan
Because (1) we don't necessarily trust it (may be wrong upstream, or
using a different GHC install locally), and (2) custom snapshots don't
want to provide this.

This patch is not complete on its own. In particular, it's possible
(such as with the script command) to try to load up the global package
information before GHC is installed. Next step is to have a proper
separation between a resolved and unresolved build plan.