Upgrade stack to the latest version of hpack
hpack recently updated to version 0.18 and one of the big changes from
that is the ability to handle Happy and Alex files. Previously this made
using hpack with projects that contained these files impossible. As such
they were stuck with using cabal files.

This commit updates stack to this newer version of hpack so that stack
can be used with hpack or cabal based projects interchangeably if Alex
or Happy files are involved.