Switch from Local/Upstream to Files/Index
This is bringing us closer to the right way of handling things, now that
extensible snapshots are a reality. We have two kinds of packages: those
with files on the filesystem, and those with files from an index. We
need to break the idea that only local packages depend on the
filesystem, it's no longer true.