Do less EnvOverride explicit passing #3607
See
https://github.com/commercialhaskell/stack/issues/3607#issuecomment-347183460.
The basic idea is: if we pass around EnvOverride information in both the
implicit Config and an explicit argument, we're more likely to forget to
update one of them, which in fact happened in #3607. By removing the
explicit argument, it becomes more obvious when we're not using an
updated EnvOverride.

As a side effect: I think this code is easier to read with less argument
overhead.