Don't track hpack run cache separately
Instead: just cache the results of cabal file parsing, and run hpack
when doing so. This (as the previous few patches) involved much more
overhaul than seems like it should. The best way to do this reliably is
to only expose a single function from Stack.Package which can run hpack.
In turn, this ended up requiring a conversion of a bunch of parts of the
code base from passing around Path Abs File (pointing to the cabal file
itself) to instead pass around Path Abs Dir (pointing to the directory).

I think this is a good change, once against simplifying things a bit
more.