More compat with buggy Cabal buildable deps #3631
This is a direct continuation of my (now realized to be incorrect) patch
in 2a52ac28d28fda0b637cbb09c3c74d1de5605a1f. This ensures that tests and
benchmarks are only switched to `buildable = True` if we're enabling
them, otherwise we can end up with cyclic dependency errors.