Handle internal libraries when building package description for GHCi.
When GHCi starts it looks for a package in the format given by Cabal but
stack used to pass the package name as the name of the sublibrary. So we
remove the sublibrary from the list of packages GHCi should search for
and add the munged name to the same list by using the `omitPkgs` and
`addPkgs` lists.