Possible fix for #3992
See #3992. This adds, to `main`, an initial output of the 'SGR Reset' ANSI code to `stdout`, if `stdout` is recognised as a supported terminal by `hSupportsANSI`. For a native (ConHost) terminal on Windows 10, this will enable its ANSI-capability.

The `ColorNever` option is respected. To do so, type `ColorWhen` is made an instance of `Eq`.