Windows: use hIsTerminalDevice or isMinTTYHandle
'mintty' is a terminal emulator for 'Cygwin', 'MSYS' or 'MSYS2', and dervied projects, and for 'WSL' (Windows Subsystem for Linux). `hIsTerminalDevice` does not recognise handles to 'mintty' terminals, but `isMinTTYHandle` does. Currently, `stack` relies on `hIsTerminalDevice` in three places to detect handles to terminals and does not use `isMinTTYHandle`.

`isMinTTYHandle` was first added to package `Win32-2.5.0.0` and then made robust to an upstream (ntdll-related) bug from `Win32-2.5.3.0`. The latter constraint is added to `package.yaml`.