Re #4091: Pre-Windows 10, default --color=never
Implements `--color=never` as default on Windows before Windows 10, if emulation would be required to support ANSI colour codes.

Also updates `stack.yaml`, in order to benefit from `ansi-terminal-0.8.0.4`.

`defaultColorWhen`, exported by new module `Stack.DefaultColorWhen`, establshes the default. That module differs, pending on whether the operating system is Windows or non-Windows (see the conditional in `package.yaml`).

`globalOptsFromMonoid` in module `Stack.Options.GlobalParser` is modified so that the default is an argument rather than a given.

A beneficial side-effect of `defaultColorWhen` on Windows is that it enables ANSI on ANSI-capable native terminals. This allows some existing code in module `Main` to be removed.