Fix issues with stderr being inherited by child processes
The only case where stdio should be inherited is when passing execution to
another process, so there should be no uses of runProcess

Most processes should also not inherit stdin, but I'm not sure if there are any
concrete issues with this in stack. I recall some prior issues where some
processes do not like being given a closed stdin, for example.

There is one remaining use of readProcessStdout_ in Stack.Docker which should
probably be fixed. However, I'm not sure what the correct implementation there
should be. The comment in the code makes it seem like it should log stderr while
reading stdout.