Fix #4183 Make PrettyPrint styles background-agnostic
In order to ensure sufficient contrast on both black (dull black) and white (vivid white) terminals, also changes foreground text `bold . white` to `dullcyan` (affecting `styleFile` and `styleUrl`) and `yellow` to `dullyellow` (affecting `styleWarning` and `styleCurrent`).

`dullcyan` is used in preference to `black` (vivid black) because the latter is problematic with the solarised dark theme.

See also issue #4047 and (reverted) #4814.