Do not hardcode GHC versions for Nixpkgs
Hardcoding GHC versions for Nixpkgs makes assumptions which are frequently
violated: that the user has the latest Nixpkgs and that the user is using the
latest Stack. Instead of hardcoding the latest minor version in each major
branch, the latest minor version in the user's Nixpkgs is selected for the
wanted regular branch.