Fix #4205 (Step 0): Annotate with Style, not [SGR]
These changes do not affect the existing functionality of stack but provide a foundation for the next steps of suggestion #4205.

Stack's styles are implemented as type `Style` in new module `Stack.PrettyPrint`, which also defines a `StyleSpec` as `(Text, [SGR])` (the first item providing the name of the key) and `Styles` as `Data.Array.IArray.Array Style StyleSpec`. Package `array` is added as a dependency in `package.yaml`, accordingly.

Given the use of `cyan` in the original instance of `Display` for `(PackageName, NamedComponent)`, a new _style_ `PkgComponent` (with key `package-component`) is added to `Style`.

Given the use of `green` in `displayMilliseconds`, it is assumed the intended style was `Good`.

Functions like `StyleWarning`, `StyleError` can be, and are, replaced throughout by `style :: Style -> ...` (eg `style Warning`, `style Error`).

The type `Runner` is modified to include field `runnerStyles :: !Styles`. New module `Stack.DefaultStyles` provides `defaultStyles`. Consequent changes are made to the plumbing, including the tests.

`AnsiDoc` is replaced by `type StyleDoc = Doc StyleAnn`, throughout. In module `Stack.PrettyPrint`, use is made of the `HasRunner` environment and it is there that `SimpleDoc StyleDoc` is translated to `SimpleDoc AnsiDoc`.