Fix #4205 (Step 5): Add `stack ls stack-colors`
This builds on the pull request for step 0. It adds `stack ls stack-colors` to the `stack ls` command, in module `Stack.Ls`. The command's functionality is provided by `listStylesCmd`. The rest of the changes add the `optparse-applicative` plumbing.

The `ls` command documentation, in module `Main`, is updated.

`ChangeLog.md` is updated.