Fix #4205 (Step 3): Add `--stack-colors=STYLES`
This builds on the pull request for step 0. It adds the `--stack-colors=STYLES` global option. The functionality is primarily implemented via `readStyles` in module `Stack.Config` and `parseStylesUpdateFromString` in new module `Stack.Types.StylesUpdate`. `GlobalOptsMonoid` is extended to include `globalMonoidStyles`; a list of `(Style, SpecStyle)` that are then used to update the array that is `defaultStyles`.

24-bit colour is now a possibility and so package `colour` is added as a dependency in `package.yaml`.

`ChangeLog.md` is updated, accordingly.